save_image_service/requirements.txt
fastapi==0.95.2
uvicorn==0.22.0
gunicorn==20.1.0
ydb[yc]==3.5.0
aioboto3==11.1.0
python-multipart==0.0.6
python-dotenv==1.0.0
grpcio==1.60.0
protobuf==4.25.1
apscheduler==3.10.1